refactor: Moved `print_gcs_document_tree()` to utilities. - Added `print_gcs_document_tree()` to `quickstart_sample`
