+++ samples/snippets/test_quickstart_sample.py
@@ -30,5 +30,7 @@ def test_quickstart_sample(capsys: pytest.CaptureFixture) -> None:
     )
     out, _ = capsys.readouterr()
 
+    assert "Document structure in Cloud Storage" in out
+    assert gcs_bucket_name in out
     assert "Number of Pages: 1" in out
     assert "Number of Entities: 35" in out
