+++ tests/unit/test_document.py
@@ -28,8 +28,6 @@ import glob
 from google.cloud.documentai_toolbox import document
 
 from google.cloud import documentai
-from google.cloud import storage
-
 from google.cloud.vision import AnnotateFileResponse
 
 
@@ -189,150 +187,6 @@ def test_document_from_gcs_with_unordered_shards(get_bytes_unordered_files_mock)
 
     for page_index, page in enumerate(actual.pages):
         assert page.documentai_page.page_number == page_index + 1
-
-
-@mock.patch("google.cloud.documentai_toolbox.wrappers.document.storage")
-def test_print_gcs_document_tree_with_one_folder(mock_storage, capfd):
-    client = mock_storage.Client.return_value
-
-    mock_bucket = mock.Mock()
-
-    client.Bucket.return_value = mock_bucket
-
-    blobs = [
-        storage.Blob(
-            name="gs://test-directory/1/test_shard1.json",
-            bucket="gs://test-directory/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/1/test_shard2.json",
-            bucket="gs://test-directory/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/1/test_shard3.json",
-            bucket="gs://test-directory/1",
-        ),
-    ]
-
-    client.list_blobs.return_value = blobs
-
-    document.print_gcs_document_tree(gcs_bucket_name="test-directory", gcs_prefix="/")
-
-    mock_storage.Client.assert_called_once()
-
-    out, err = capfd.readouterr()
-    assert (
-        out
-        == """gs://test-directory/1
-├──test_shard1.json
-├──test_shard2.json
-└──test_shard3.json\n\n"""
-    )
-
-
-@mock.patch("google.cloud.documentai_toolbox.wrappers.document.storage")
-def test_print_gcs_document_tree_with_3_documents(mock_storage, capfd):
-    client = mock_storage.Client.return_value
-
-    mock_bucket = mock.Mock()
-
-    client.Bucket.return_value = mock_bucket
-
-    blobs = [
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard1.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard2.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard3.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-    ]
-
-    client.list_blobs.return_value = blobs
-
-    document.print_gcs_document_tree(
-        gcs_bucket_name="test-directory", gcs_prefix="documentai/output/123456789/1/"
-    )
-
-    mock_storage.Client.assert_called_once()
-
-    out, err = capfd.readouterr()
-    assert (
-        out
-        == """gs://test-directory/documentai/output/123456789/1
-├──test_shard1.json
-├──test_shard2.json
-└──test_shard3.json\n\n"""
-    )
-
-
-@mock.patch("google.cloud.documentai_toolbox.wrappers.document.storage")
-def test_print_gcs_document_tree_with_more_than_5_document(mock_storage, capfd):
-    client = mock_storage.Client.return_value
-
-    mock_bucket = mock.Mock()
-
-    client.Bucket.return_value = mock_bucket
-
-    blobs = [
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard1.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard2.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard3.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard4.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard5.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-        storage.Blob(
-            name="gs://test-directory/documentai/output/123456789/1/test_shard6.json",
-            bucket="gs://test-directory/documentai/output/123456789/1",
-        ),
-    ]
-    client.list_blobs.return_value = blobs
-
-    document.print_gcs_document_tree(
-        gcs_bucket_name="test-directory", gcs_prefix="documentai/output/123456789/1/"
-    )
-
-    mock_storage.Client.assert_called_once()
-
-    out, err = capfd.readouterr()
-    assert (
-        out
-        == """gs://test-directory/documentai/output/123456789/1
-├──test_shard1.json
-├──test_shard2.json
-├──test_shard3.json
-├──test_shard4.json
-├──test_shard5.json
-│  ....
-└──test_shard6.json\n\n"""
-    )
-
-
-def test_print_gcs_document_tree_with_gcs_uri_contains_file_type():
-    with pytest.raises(ValueError, match="gcs_prefix cannot contain file types"):
-        document.print_gcs_document_tree(
-            gcs_bucket_name="test-directory",
-            gcs_prefix="documentai/output/123456789/1/test_file.json",
-        )
 
 
 def test_search_page_with_target_string(get_bytes_single_file_mock):
